# docs/openapi.yaml
{
  "openapi": "3.0.0",
  "info": {
    "title": "LigueLead API",
    "version": "1.0.0",
    "description": "Sistema completo de gerenciamento de projetos e tarefas com cache Redis profissional",
    "contact": {
      "name": "Suporte LigueLead",
      "email": "suporte@liguelead.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desenvolvimento"
    }
  ],
...
  "paths": {
...
    "/projects/{projectId}/tasks": {
      "post": {
        "summary": "Criar nova tarefa em um projeto",
        "tags": [
          "Tarefas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do projeto",
            "example": 1
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "pt",
                "en",
                "es"
              ],
              "default": "pt"
            },
            "description": "Idioma da resposta (pt=Português, en=English, es=Español)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Título da tarefa",
                    "example": "Implementar autenticação"
                  },
                  "description": {
                    "type": "string",
                    "description": "Descrição da tarefa",
                    "example": "Desenvolver sistema de login e registro"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "in_progress",
                      "completed",
                      "cancelled"
                    ],
                    "default": "pending",
                    "description": "Status da tarefa"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "low",
                      "medium",
                      "high",
                      "urgent"
                    ],
                    "default": "medium",
                    "description": "Prioridade da tarefa"
                  },
                  "dueDate": {
                    "type": "string",
                    "format": "date",
                    "description": "Data de vencimento",
                    "example": "2024-02-15"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarefa criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "title": {
                      "type": "string",
                      "example": "Implementar autenticação"
                    },
                    "description": {
                      "type": "string",
                      "example": "Desenvolver sistema de login e registro"
                    },
                    "status": {
                      "type": "string",
                      "example": "pending"
                    },
                    "priority": {
                      "type": "string",
                      "example": "high"
                    },
                    "dueDate": {
                      "type": "string",
                      "format": "date",
                      "example": "2024-02-15"
                    },
                    "projectId": {
                      "type": "integer",
                      "example": 1
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Projeto não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "get": {
        "summary": "Listar tarefas de um projeto",
        "tags": [
          "Tarefas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do projeto",
            "example": 1
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "pt",
                "en",
                "es"
              ],
              "default": "pt"
            },
            "description": "Idioma da resposta (pt=Português, en=English, es=Español)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tarefas do projeto retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "title": {
                        "type": "string",
                        "example": "Implementar autenticação"
                      },
                      "description": {
                        "type": "string",
                        "example": "Desenvolver sistema de login e registro"
                      },
                      "status": {
                        "type": "string",
                        "example": "pending"
                      },
                      "priority": {
                        "type": "string",
                        "example": "high"
                      },
                      "dueDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-15"
                      },
                      "projectId": {
                        "type": "integer",
                        "example": 1
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Projeto não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Listar todas as tarefas",
        "tags": [
          "Tarefas"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "pt",
                "en",
                "es"
              ],
              "default": "pt"
            },
            "description": "Idioma da resposta (pt=Português, en=English, es=Español)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tarefas retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "title": {
                        "type": "string",
                        "example": "Implementar autenticação"
                      },
                      "description": {
                        "type": "string",
                        "example": "Desenvolver sistema de login e registro"
                      },
                      "status": {
                        "type": "string",
                        "example": "pending"
                      },
                      "priority": {
                        "type": "string",
                        "example": "high"
                      },
                      "dueDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-15"
                      },
                      "projectId": {
                        "type": "integer",
                        "example": 1
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
  }
}